requirements.txt

fastapi
uvicorn
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
sqlalchemy
python-telegram-bot
httpx
